Improve file management for tests
