+++ src/mobile_app/test/account_setup/fixtures/fixtures.dart
@@ -1,0 +1,2 @@
+export 'name_form_fixture.dart';
+export 'nick_name_form_fixture.dart';
