+++ src/mobile_app/test/account_setup/widgets/nick_name_form_screen_test.dart
@@ -1,15 +1,15 @@
 import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:formz/formz.dart';
 import 'package:mobile_app/account_setup/blocs/sign_up/sign_up_bloc.dart';
 import 'package:mobile_app/account_setup/models/models.dart';
 import 'package:mobile_app/account_setup/models/name_field.dart';
-import 'package:mobile_app/account_setup/widgets/nick_name_form.dart';
 import 'package:mobile_app/account_setup/widgets/nick_name_form_screen.dart';
 import 'package:mobile_app/positive_affirmations_keys.dart';
 import 'package:mocktail/mocktail.dart';
+
+import '../fixtures/fixtures.dart';
 
 class FakeSignUpEvent extends Fake implements SignUpEvent {}
 
@@ -19,24 +19,6 @@ class MockSignUpBloc extends MockBloc<SignUpEvent, SignUpState>
     implements SignUpBloc {}
 
 class MockNickNameField extends Mock implements NameField {}
-
-class NickNameFormFixture extends StatelessWidget {
-  NickNameFormFixture(this.bloc);
-
-  final SignUpBloc bloc;
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocProvider.value(
-      value: bloc,
-      child: MaterialApp(
-        home: Scaffold(
-          body: NickNameForm(),
-        ),
-      ),
-    );
-  }
-}
 
 void main() {
   const mockValidName = 'validName';
